Move FormConfirmDialog to Dialog directory
